+++ src/AspireApp/AppHost.cs
@@ -1,37 +1,36 @@
+using Aspire.Hosting.ApplicationModel;
 using Projects;
 
 var builder = DistributedApplication.CreateBuilder(args);
+var javaAgentPath = Path.Combine(builder.AppHostDirectory, "agents", "opentelemetry-javaagent.jar");
+var aspireCertificatePath = Path.Combine(builder.AppHostDirectory, "agents", "aspire-localhost-cert.pem");
 
 var rabbitUser = builder.AddParameter("rabbitmq-user", "guest", secret: true);
 var rabbitPassword = builder.AddParameter("rabbitmq-password", "guest", secret: true);
 
-var rabbitmq = builder.AddRabbitMQ("messaging", rabbitUser, rabbitPassword, port: 5672)
-    .WithManagementPlugin(port: 15672)
-    .WithDataVolume(isReadOnly: false);
+var rabbitmq = builder.AddRabbitMQ("messaging", rabbitUser, rabbitPassword)
+    .WithManagementPlugin()
+    .WithImageTag("4.1.8-management-alpine");
 
 var csharpTestApp = builder.AddProject<TestApp>("testapp")
     .WithReference(rabbitmq)
-    .WithEndpoint("https", endpoint =>
-    {
-        endpoint.Port = 7244;
-        endpoint.IsExternal = true;
-    })
-    .WithEndpoint("http", endpoint =>
-    {
-        endpoint.Port = 5112;
-        endpoint.IsExternal = true;
-    })
+    .WithEnvironment("RABBITMQ_HOST", rabbitmq.Resource.PrimaryEndpoint.Property(EndpointProperty.Host))
+    .WithEnvironment("RABBITMQ_PORT", rabbitmq.Resource.PrimaryEndpoint.Property(EndpointProperty.Port))
+    .WithExternalHttpEndpoints()
     .WaitFor(rabbitmq);
 
-var javaTestApp = builder.AddJavaApp(
+var javaTestApp = builder.AddExecutable(
     "testapp-java",
-    workingDirectory: "../Java/testapp",
-    new JavaAppExecutableResourceOptions
-    {
-        ApplicationName = "build/libs/testapp-1.0-SNAPSHOT.jar",
-        OtelAgentPath = "../../AspireApp/agents",
-    })
-    .WithEnvironment("OTEL_EXPORTER_OTLP_CERTIFICATE", "../../AspireApp/agents/aspire-localhost-cert.pem")
+    "gradle",
+    workingDirectory: "../..",
+    ":testapp:run",
+    "--no-daemon")
+    .WithOtelAgent(javaAgentPath)
+    .WithHttpEndpoint(targetPort: 5301, name: "http")
+    .WithEnvironment("OTEL_EXPORTER_OTLP_CERTIFICATE", aspireCertificatePath)
+    .WithEnvironment("HTTP_PORT", "5301")
+    .WithEnvironment("RABBITMQ_HOST", rabbitmq.Resource.PrimaryEndpoint.Property(EndpointProperty.Host))
+    .WithEnvironment("RABBITMQ_PORT", rabbitmq.Resource.PrimaryEndpoint.Property(EndpointProperty.Port))
     .WithReference(rabbitmq)
     .WithExternalHttpEndpoints()
     .WaitFor(rabbitmq);